refactor: Move TargetingBuilderHeader inside TargetingBuilder
