+++ src/features/EmailMarketing/shared/TargetingBuilder/components/TargetingBuilderHeader/TargetingBuilderHeader.scss
