+++ src/features/EmailMarketing/shared/TargetingBuilder/components/TargetingBuilderHeader/components/AIAccuracyWarning/AIAccuracyWarning.test.tsx
@@ -1,9 +1,9 @@
 import {fireEvent, render, screen} from '@testing-library/react';
 import AIAccuracyWarning from './AIAccuracyWarning';
-import {emailMarketingInitialState} from '../../../../store/emailMarketingInitialState';
-import {renderComponentWithState} from '../../../../store/emailMarketingStoreUtils';
-import {EmailMarketingActionType} from '../../../../store/emailMarketingActionTypes';
-import {mockDispatch} from '../../../../../../utils/mocks/mocks';
+import {emailMarketingInitialState} from '../../../../../../store/emailMarketingInitialState';
+import {renderComponentWithState} from '../../../../../../store/emailMarketingStoreUtils';
+import {EmailMarketingActionType} from '../../../../../../store/emailMarketingActionTypes';
+import {mockDispatch} from '../../../../../../../../utils/mocks/mocks';
 
 describe('AIAccuracyWarning', () => {
   it('renders the component', async () => {
