+++ src/features/EmailMarketing/shared/TargetingBuilder/components/TargetingBuilderHeader/components/AIErrorWarnings/AIErrorWarnings.test.tsx
@@ -1,9 +1,9 @@
 import {fireEvent, render, screen} from '@testing-library/react';
 import AIErrorWarnings from './AIErrorWarnings';
-import {EmailMarketingActionType} from '../../../../store/emailMarketingActionTypes';
-import {mockDispatch} from '../../../../../../utils/mocks/mocks';
-import {renderComponentWithState} from '../../../../store/emailMarketingStoreUtils';
-import {emailMarketingInitialState} from '../../../../store/emailMarketingInitialState';
+import {EmailMarketingActionType} from '../../../../../../store/emailMarketingActionTypes';
+import {mockDispatch} from '../../../../../../../../utils/mocks/mocks';
+import {renderComponentWithState} from '../../../../../../store/emailMarketingStoreUtils';
+import {emailMarketingInitialState} from '../../../../../../store/emailMarketingInitialState';
 
 describe('AIErrorWarnings', () => {
   const mockErrors = [
