+++ src/features/EmailMarketing/shared/TargetingBuilder/components/TargetingBuilderHeader/components/AIErrorWarnings/AIErrorWarnings.tsx
@@ -1,7 +1,7 @@
-import Warning from '../../../../../shared/ui/components/Warning/Warning';
-import {EmailMarketingActionType} from '../../../../store/emailMarketingActionTypes';
-import {useEmailMarketingState} from '../../../../store/useEmailMarketingState';
-import {ErrorWarning} from '../../../emailMarketingTypes';
+import Warning from '../../../../../../../shared/ui/components/Warning/Warning';
+import {EmailMarketingActionType} from '../../../../../../store/emailMarketingActionTypes';
+import {useEmailMarketingState} from '../../../../../../store/useEmailMarketingState';
+import {ErrorWarning} from '../../../../../emailMarketingTypes';
 import './AIErrorWarnings.scss';
 
 const AIErrorWarnings = () => {
